create figures directory for separate plotting back ends
