+++ solarforecastarbiter/reports/figures/plotly_figures.py
